+++ crates/swc/src/lib.rs
@@ -136,7 +136,7 @@ use swc_common::{
 pub use swc_compiler_base::{PrintArgs, TransformOutput};
 pub use swc_config::config_types::{BoolConfig, BoolOr, BoolOrDataConfig};
 use swc_ecma_ast::{EsVersion, Program};
-use swc_ecma_codegen::Node;
+use swc_ecma_codegen::{to_code, Node};
 use swc_ecma_loader::resolvers::{
     lru::CachingResolver, node::NodeModulesResolver, tsc::TsConfigResolver,
 };
@@ -155,6 +155,8 @@ use swc_ecma_visit::{FoldWith, VisitMutWith, VisitWith};
 pub use swc_error_reporters::handler::{try_with_handler, HandlerOpts};
 pub use swc_node_comments::SwcComments;
 use swc_timer::timer;
+use swc_transform_common::output::emit;
+use swc_typescript::fast_dts::FastDts;
 use tracing::warn;
 use url::Url;
 
@@ -731,7 +733,7 @@ impl Compiler {
                 None
             };
 
-            self.apply_transforms(handler, orig, config)
+            self.apply_transforms(handler, fm.clone(), orig, config)
         })
     }
 
@@ -949,11 +951,20 @@ impl Compiler {
     fn apply_transforms(
         &self,
         handler: &Handler,
+        fm: Arc<SourceFile>,
         orig: Option<sourcemap::SourceMap>,
         config: BuiltInput<impl swc_ecma_visit::Fold>,
     ) -> Result<TransformOutput, Error> {
         self.run(|| {
             let program = config.program;
+
+            if config.emit_isolated_dts && !config.syntax.typescript() {
+                handler.warn(
+                    "jsc.experimental.emitIsolatedDts is enabled but the syntax is not TypeScript",
+                );
+            }
+
+            let emit_dts = config.syntax.typescript() && config.emit_isolated_dts;
             let source_map_names = if config.source_maps.enabled() {
                 let mut v = swc_compiler_base::IdentCollector {
                     names: Default::default(),
@@ -966,8 +977,34 @@ impl Compiler {
                 Default::default()
             };
 
+            let dts_code = if emit_dts && program.is_module() {
+                let mut checker = FastDts::new(fm.name.clone().into());
+                let mut module = program.clone().expect_module();
+
+                let issues = checker.transform(&mut module);
+
+                for issue in issues {
+                    let range = issue.range();
+
+                    handler
+                        .struct_span_err(range.span, &issue.to_string())
+                        .emit();
+                }
+                let dts_code = to_code(&module);
+                Some(dts_code)
+            } else {
+                None
+            };
+
             let mut pass = config.pass;
             let (program, output) = swc_transform_common::output::capture(|| {
+                if let Some(dts_code) = dts_code {
+                    emit(
+                        "__swc_isolated_declarations__".into(),
+                        serde_json::Value::String(dts_code),
+                    );
+                }
+
                 helpers::HELPERS.set(&Helpers::new(config.external_helpers), || {
                     HANDLER.set(handler, || {
                         // Fold module