+++ crates/swc/tests/ts-isolated-declaration/oxc/input/arrow-function-return-type.ts
@@ -1,0 +1,9 @@
+function A() {
+  return () => {
+    return C;
+  }
+}
+
+const B = () => { return B };
+
+const C = function () {}