+++ crates/swc/tests/ts-isolated-declaration/oxc/input/generator.ts
@@ -1,0 +1,10 @@
+// Correct
+function *generatorGood(): Generator<number> {}
+
+
+// Need to explicit return type for async functions
+// Incorrect
+function *generatorGoodBad() {
+  yield 50;
+  return 42;
+}