+++ crates/swc/tests/ts-isolated-declaration/oxc/output/empty-export.d.ts
@@ -1,0 +1,2 @@
+type A = string;
+export declare function a(): A;