+++ crates/swc/tests/ts-isolated-declaration/oxc/output/infer-template-literal.ts
@@ -1,0 +1,11 @@
+export var CSS_VARS_HELPER = "useCssVars";
+export function g() {
+    var func = arguments.length > 0 && arguments[0] !== void 0 ? arguments[0] : "useCssVar";
+}
+export var F = {
+    a: "a",
+    b: [
+        "b"
+    ]
+};
+export var BAD = "useCssV".concat(v, "ars");