+++ crates/swc/tests/ts-isolated-declaration/oxc/output/readonly.ts
@@ -1,0 +1,2 @@
+export var EMPTY_OBJ = __DEV__ ? Object.freeze({}) : {};
+export var EMPTY_ARR = __DEV__ ? Object.freeze([]) : [];